Extracted dependency dumping into the new Tracer class
Kept the Annotations::PrepareDependenciesDump methods, as they are
still needed for the mobile linkers. They will be removed once we
update them.
