+++ linker/Linker.Steps/ResolveFromXmlStep.cs
@@ -110,7 +110,7 @@ namespace Mono.Linker.Steps {
 
 		protected void ProcessAssembly (AssemblyDefinition assembly, XPathNodeIterator iterator)
 		{
-			Annotations.Push (assembly);
+			Tracer.Push (assembly);
 			if (GetTypePreserve (iterator.Current) == TypePreserve.All) {
 				foreach (var type in assembly.MainModule.Types)
 					MarkAndPreserveAll (type);
@@ -118,7 +118,7 @@ namespace Mono.Linker.Steps {
 				ProcessTypes (assembly, iterator.Current.SelectChildren ("type", _ns));
 				ProcessNamespaces (assembly, iterator.Current.SelectChildren ("namespace", _ns));
 			}
-			Annotations.Pop ();
+			Tracer.Pop ();
 		}
 
 		void ProcessNamespaces (AssemblyDefinition assembly, XPathNodeIterator iterator)
@@ -140,14 +140,14 @@ namespace Mono.Linker.Steps {
 			Annotations.SetPreserve (type, TypePreserve.All);
 
 			if (!type.HasNestedTypes) {
-				Annotations.Pop ();
+				Tracer.Pop ();
 				return;
 			}
 
 			foreach (TypeDefinition nested in type.NestedTypes)
 				MarkAndPreserveAll (nested);
 
-			Annotations.Pop ();
+			Tracer.Pop ();
 		}
 
 		void ProcessTypes (AssemblyDefinition assembly, XPathNodeIterator iterator)
@@ -167,10 +167,10 @@ namespace Mono.Linker.Steps {
 					if (assembly.MainModule.HasExportedTypes) {
 						foreach (var exported in assembly.MainModule.ExportedTypes) {
 							if (fullname == exported.FullName) {
-								Annotations.Push (exported);
+								Tracer.Push (exported);
 								MarkingHelpers.MarkExportedType (exported, assembly.MainModule);
 								var resolvedExternal = exported.Resolve ();
-								Annotations.Pop ();
+								Tracer.Pop ();
 								if (resolvedExternal != null) {
 									type = resolvedExternal;
 									break;
@@ -251,7 +251,7 @@ namespace Mono.Linker.Steps {
 			}
 
 			Annotations.MarkAndPush (type);
-			Annotations.AddDirectDependency (this, type);
+			Tracer.AddDirectDependency (this, type);
 
 			if (type.IsNested) {
 				var parent = type;
@@ -270,7 +270,7 @@ namespace Mono.Linker.Steps {
 				MarkSelectedEvents (nav, type);
 				MarkSelectedProperties (nav, type);
 			}
-			Annotations.Pop ();
+			Tracer.Pop ();
 		}
 
 		void MarkSelectedFields (XPathNavigator nav, TypeDefinition type)
@@ -409,7 +409,7 @@ namespace Mono.Linker.Steps {
 		void MarkMethod (MethodDefinition method)
 		{
 			Annotations.Mark (method);
-			Annotations.AddDirectDependency (this, method);
+			Tracer.AddDirectDependency (this, method);
 			Annotations.SetAction (method, MethodAction.Parse);
 		}
